geo_traits

Trait CoordTrait

pub trait CoordTrait {
    type T;

    // Required methods
    fn dim(&self) -> Dimensions;
    fn x(&self) -> Self::T;
    fn y(&self) -> Self::T;
    fn nth_or_panic(&self, n: usize) -> Self::T;

    // Provided methods
    fn nth(&self, n: usize) -> Option<Self::T> { ... }
    fn x_y(&self) -> (Self::T, Self::T) { ... }
    unsafe fn nth_unchecked(&self, n: usize) -> Self::T { ... }
}

A trait for accessing data from a generic Coord.

Refer to [geo_types::Coord] for information about semantics and validity.

Required Associated Types

type T

The coordinate type of this geometry

Required Methods

fn dim(&self) -> Dimensions

Dimensions of the coordinate tuple

fn x(&self) -> Self::T

x component of this coord.

fn y(&self) -> Self::T

y component of this coord.

fn nth_or_panic(&self, n: usize) -> Self::T

Access the n’th (0-based) element of the CoordinateTuple. May panic if n >= DIMENSION. See also nth().

Provided Methods

fn nth(&self, n: usize) -> Option<Self::T>

Access the n’th (0-based) element of the CoordinateTuple. Returns None if n >= DIMENSION.

See also nth_or_panic() and nth_unchecked().

Panics

This method may panic if dim() does not correspond to the actual number of dimensions in this coordinate.

fn x_y(&self) -> (Self::T, Self::T)

Returns a tuple that contains the x/horizontal & y/vertical component of the coord.

unsafe fn nth_unchecked(&self, n: usize) -> Self::T

Access the n’th (0-based) element of the CoordinateTuple. May panic if n >= DIMENSION.

See also nth(), nth_or_panic().

You might want to override the default implementation of this method if you can provide a more efficient implementation.

Safety

Though it may panic, the default implementation actually is safe. However, implementors are allowed to implement this method themselves with an unsafe implementation. See the individual implementations for more information on their own Safety considerations.