geo_traits

Trait MultiLineStringTrait

pub trait MultiLineStringTrait: Sized + GeometryTrait {
    type InnerLineStringType<'a>: 'a + LineStringTrait<T = Self::T>
       where Self: 'a;

    // Required methods
    fn num_line_strings(&self) -> usize;
    unsafe fn line_string_unchecked(
        &self,
        i: usize,
    ) -> Self::InnerLineStringType<'_>;

    // Provided methods
    fn line_strings(
        &self,
    ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::InnerLineStringType<'_>> { ... }
    fn line_string(&self, i: usize) -> Option<Self::InnerLineStringType<'_>> { ... }
}

A trait for accessing data from a generic MultiLineString.

A MultiLineString is a collection of LineStrings.

Refer to [geo_types::MultiLineString] for information about semantics and validity.

Required Associated Types

type InnerLineStringType<'a>: 'a + LineStringTrait<T = Self::T> where Self: 'a

The type of each underlying LineString, which implements LineStringTrait

Required Methods

fn num_line_strings(&self) -> usize

The number of line_strings in this MultiLineString

unsafe fn line_string_unchecked( &self, i: usize, ) -> Self::InnerLineStringType<'_>

Access to a specified line_string in this MultiLineString

Safety

Accessing an index out of bounds is UB.

Provided Methods

fn line_strings( &self, ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::InnerLineStringType<'_>>

An iterator over the LineStrings in this MultiLineString

fn line_string(&self, i: usize) -> Option<Self::InnerLineStringType<'_>>

Access to a specified line_string in this MultiLineString Will return None if the provided index is out of bounds

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.