geo_traits

Trait MultiPointTrait

pub trait MultiPointTrait: Sized + GeometryTrait {
    type InnerPointType<'a>: 'a + PointTrait<T = Self::T>
       where Self: 'a;

    // Required methods
    fn num_points(&self) -> usize;
    unsafe fn point_unchecked(&self, i: usize) -> Self::InnerPointType<'_>;

    // Provided methods
    fn points(
        &self,
    ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::InnerPointType<'_>> { ... }
    fn point(&self, i: usize) -> Option<Self::InnerPointType<'_>> { ... }
}

A trait for accessing data from a generic MultiPoint.

A MultiPoint is a collection of Points.

Refer to [geo_types::MultiPoint] for information about semantics and validity.

Required Associated Types

type InnerPointType<'a>: 'a + PointTrait<T = Self::T> where Self: 'a

The type of each underlying Point, which implements PointTrait

Required Methods

fn num_points(&self) -> usize

The number of points in this MultiPoint

unsafe fn point_unchecked(&self, i: usize) -> Self::InnerPointType<'_>

Access to a specified point in this MultiPoint

Safety

Accessing an index out of bounds is UB.

Provided Methods

fn points( &self, ) -> impl DoubleEndedIterator + ExactSizeIterator<Item = Self::InnerPointType<'_>>

An iterator over the points in this MultiPoint

fn point(&self, i: usize) -> Option<Self::InnerPointType<'_>>

Access to a specified point in this MultiPoint Will return None if the provided index is out of bounds

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.