geo_traits

Trait TriangleTrait

pub trait TriangleTrait: Sized + GeometryTrait {
    type CoordType<'a>: 'a + CoordTrait<T = Self::T>
       where Self: 'a;

    // Required methods
    fn first(&self) -> Self::CoordType<'_>;
    fn second(&self) -> Self::CoordType<'_>;
    fn third(&self) -> Self::CoordType<'_>;

    // Provided method
    fn coords(&self) -> [Self::CoordType<'_>; 3] { ... }
}

A trait for accessing data from a generic Triangle.

A triangle is a bounded area whose three vertices are defined by coordinates.

Refer to [geo_types::Triangle] for information about semantics and validity.

Required Associated Types

type CoordType<'a>: 'a + CoordTrait<T = Self::T> where Self: 'a

The type of each underlying coordinate, which implements CoordTrait

Required Methods

fn first(&self) -> Self::CoordType<'_>

Access the first coordinate in this Triangle

fn second(&self) -> Self::CoordType<'_>

Access the second coordinate in this Triangle

fn third(&self) -> Self::CoordType<'_>

Access the third coordinate in this Triangle

Provided Methods

fn coords(&self) -> [Self::CoordType<'_>; 3]

Access the three underlying coordinates

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.